requirements.txt
cx_Oracle
psycopg2
sqlalchemy
pandas
apache-airflow